REFACTOR: moving timer definitions to MCU locations.
