+++ make/mcu/AT32F4.mk
@@ -1,0 +1,6 @@
+#
+# AT32F4 Make file include
+#
+
+INCLUDE_DIRS    := $(INCLUDE_DIRS) \
+                   $(ROOT)/src/main/drivers/at32
