+++ src/main/drivers/at32/timer_at32f43x.c
@@ -24,10 +24,9 @@
 
 #include "common/utils.h"
 
+#include "timer_def.h"
 #include "drivers/dma.h"
 #include "drivers/io.h"
-#include "drivers/timer_def.h"
-
 #include "drivers/rcc.h"
 #include "drivers/timer.h"
