+++ src/main/drivers/stm32/timer_stm32f4xx.c
@@ -26,7 +26,7 @@
 
 #include "drivers/dma.h"
 #include "drivers/io.h"
-#include "drivers/timer_def.h"
+#include "timer_def.h"
 
 #include "stm32f4xx.h"
 #include "drivers/rcc.h"
